Added Ofline to application
Rest compras and ventas of productos
AND REST SINCRONIZATION

+++ app/src/main/java/com/example/businessmanagement/vistas/Acreedor/EditarAcreedor.java
@@ -1,7 +1,10 @@
 package com.example.businessmanagement.vistas.Acreedor;
 
 import android.content.ActivityNotFoundException;
+import android.content.Context;
 import android.content.Intent;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
 import android.net.Uri;
 import android.os.Bundle;
 import android.provider.MediaStore;
@@ -17,6 +20,7 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import com.bumptech.glide.Glide;
 import com.example.businessmanagement.R;
+import com.example.businessmanagement.controladores.bdLocal.SQLAcreedoresController;
 import com.example.businessmanagement.modelos.Acreedor;
 import com.example.businessmanagement.modelos.Cliente;
 import com.example.businessmanagement.vistas.Cliente.VistaCliente;
@@ -60,7 +64,9 @@ public class EditarAcreedor extends AppCompatActivity {
         editado= false;
         aux = false;
 
-        storage = FirebaseStorage.getInstance();
+        if(isNetworkAvailable()){
+            storage = FirebaseStorage.getInstance();
+        }
 
         Bundle b = getIntent().getExtras();
 
@@ -81,50 +87,59 @@ public class EditarAcreedor extends AppCompatActivity {
     }
 
     private void cargarAcreedor() {
-        database = FirebaseDatabase.getInstance().getReference().child("Acreedores");
-        database.addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot snapshot) {
-                if(!aux) {
-                    if (editado != true) {
-                        a = snapshot.child(nif).getValue(Acreedor.class);
+        if(isNetworkAvailable()){
+            database = FirebaseDatabase.getInstance().getReference().child("Acreedores");
+            database.addValueEventListener(new ValueEventListener() {
+                @Override
+                public void onDataChange(@NonNull DataSnapshot snapshot) {
+                    if(!aux) {
+                        if (editado != true) {
+                            a = snapshot.child(nif).getValue(Acreedor.class);
 
-                        Glide.with(getApplicationContext()).load(a.getImageUri()).into(ivAcreedor);
-                        etNombre.setText(a.getNombre());
-                        etNif.setText(a.getNif());
-                        etTelefono.setText(a.getTelefono());
-                        etEmail.setText(a.getEmail());
-                    } else {
-                        for (DataSnapshot child : snapshot.getChildren()) {
-                            if (etNif.getText().toString().equals(child.child("nif").getValue().toString())) {
-                                acreedorexiste = true;
-                            }
-                        }
-                        if (acreedorexiste == false) {
-                            database = FirebaseDatabase.getInstance().getReference().child("Acreedores").child(newAcreedor.getNif());
-                            database.setValue(newAcreedor);
-                            VistaCliente.dni = newAcreedor.getNombre();
-                            Toast.makeText(getApplicationContext(), "Acreedor Modificado Correctamente", Toast.LENGTH_LONG).show();
-                            VistaCliente.tvDni.setText(newAcreedor.getNif());
-                            VistaCliente.tvNombre.setText(newAcreedor.getNombre());
-                            VistaCliente.tvTelefono.setText(newAcreedor.getTelefono());
-                            VistaCliente.tvEmail.setText(newAcreedor.getEmail());
-                            finish();
+                            Glide.with(getApplicationContext()).load(a.getImageUri()).into(ivAcreedor);
+                            etNombre.setText(a.getNombre());
+                            etNif.setText(a.getNif());
+                            etTelefono.setText(a.getTelefono());
+                            etEmail.setText(a.getEmail());
                         } else {
-                            etNombre.setText("");
-                            etNif.setText("");
-                            etEmail.setText("");
-                            etTelefono.setText("");
+                            for (DataSnapshot child : snapshot.getChildren()) {
+                                if (etNif.getText().toString().equals(child.child("nif").getValue().toString())) {
+                                    acreedorexiste = true;
+                                }
+                            }
+                            if (acreedorexiste == false) {
+                                database = FirebaseDatabase.getInstance().getReference().child("Acreedores").child(newAcreedor.getNif());
+                                database.setValue(newAcreedor);
+                                VistaCliente.dni = newAcreedor.getNombre();
+                                Toast.makeText(getApplicationContext(), "Acreedor Modificado Correctamente", Toast.LENGTH_LONG).show();
+                                VistaCliente.tvDni.setText(newAcreedor.getNif());
+                                VistaCliente.tvNombre.setText(newAcreedor.getNombre());
+                                VistaCliente.tvTelefono.setText(newAcreedor.getTelefono());
+                                VistaCliente.tvEmail.setText(newAcreedor.getEmail());
+                                finish();
+                            } else {
+                                etNombre.setText("");
+                                etNif.setText("");
+                                etEmail.setText("");
+                                etTelefono.setText("");
+                            }
                         }
                     }
                 }
-            }
 
-            @Override
-            public void onCancelled(@NonNull DatabaseError error) {
+                @Override
+                public void onCancelled(@NonNull DatabaseError error) {
 
-            }
-        });
+                }
+            });
+        }else{
+            SQLAcreedoresController sql = new SQLAcreedoresController(getApplicationContext());
+            Acreedor a = sql.getAcreedor(nif);
+            etNombre.setText(a.getNombre());
+            etNif.setText(a.getNif());
+            etTelefono.setText(a.getTelefono());
+            etEmail.setText(a.getEmail());
+        }
     }
 
     private void setup() {
@@ -135,9 +150,16 @@ public class EditarAcreedor extends AppCompatActivity {
                     Toast.makeText(getApplicationContext(),"Rellene el nif y nombre", Toast.LENGTH_LONG).show();
                 }else {
                     newAcreedor = new Acreedor(etNombre.getText().toString(), etNif.getText().toString(), etEmail.getText().toString(), etTelefono.getText().toString(), Uri.parse(a.getImageUri()), Uri.parse(a.getImgStorage()));
-                    database = FirebaseDatabase.getInstance().getReference().child("Acreedores").child(nif);
-                    editado = true;
-                    database.removeValue();
+                    SQLAcreedoresController sql = new SQLAcreedoresController(getApplicationContext());
+                    if(isNetworkAvailable()){
+                        database = FirebaseDatabase.getInstance().getReference().child("Acreedores").child(nif);
+                        editado = true;
+                        database.removeValue();
+                    }else{
+                        sql.borrarAcreedorAux(newAcreedor);
+                        sql.anadirAcreedorAux(newAcreedor);
+                    }
+                    sql.modificaAcreedor(newAcreedor);
                 }
             }
         });
@@ -152,13 +174,17 @@ public class EditarAcreedor extends AppCompatActivity {
 
     private void escogerFoto() {
 
-        Intent intent = new Intent(Intent.ACTION_PICK, MediaStore.Images.Media.INTERNAL_CONTENT_URI);
+        if(isNetworkAvailable()){
+            Intent intent = new Intent(Intent.ACTION_PICK, MediaStore.Images.Media.INTERNAL_CONTENT_URI);
 
-        try {
+            try {
 
-            startActivityForResult(intent, REQUEST_IMAGE_CAPTURE);
+                startActivityForResult(intent, REQUEST_IMAGE_CAPTURE);
 
-        } catch (ActivityNotFoundException e) {
+            } catch (ActivityNotFoundException e) {
+            }
+        }else {
+            Toast.makeText(getApplicationContext(), "Sin conexión a internet no se puede establecer una foto", Toast.LENGTH_LONG).show();
         }
 
     }
@@ -214,6 +240,13 @@ public class EditarAcreedor extends AppCompatActivity {
 
         Glide.with(getApplicationContext()).load(imageUri).into(ivAcreedor);
 
+    }
+
+    private boolean isNetworkAvailable() {
+        ConnectivityManager connectivityManager
+                = (ConnectivityManager) getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo activeNetworkInfo = connectivityManager.getActiveNetworkInfo();
+        return activeNetworkInfo != null && activeNetworkInfo.isConnected();
     }
 }
