+++ app/src/main/java/com/example/businessmanagement/vistas/Acreedor/VistaAcreedor.java
@@ -1,6 +1,9 @@
 package com.example.businessmanagement.vistas.Acreedor;
 
+import android.content.Context;
 import android.content.Intent;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -13,9 +16,8 @@ import androidx.appcompat.widget.Toolbar;
 
 import com.bumptech.glide.Glide;
 import com.example.businessmanagement.R;
+import com.example.businessmanagement.controladores.bdLocal.SQLAcreedoresController;
 import com.example.businessmanagement.modelos.Acreedor;
-import com.example.businessmanagement.modelos.Cliente;
-import com.example.businessmanagement.vistas.Cliente.EditarCliente;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
@@ -52,9 +54,7 @@ public class VistaAcreedor extends AppCompatActivity {
         ivAcreedor = findViewById(R.id.ivAcreedor);
         tbAcreedor = findViewById(R.id.tbAcreedor);
 
-        cargarCliente();
-
-
+        cargarAcreedor();
     }
 
     private void cargarToolbar(){
@@ -62,30 +62,40 @@ public class VistaAcreedor extends AppCompatActivity {
         getSupportActionBar().setTitle(acreedor.getNombre());
     }
 
-    private void cargarCliente() {
-        database = FirebaseDatabase.getInstance().getReference().child("Acreedores");
-        database.addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot snapshot) {
-                for (DataSnapshot child : snapshot.getChildren()){
-                    if(child.child("nif").getValue().toString().equals(nif)){
-                        acreedor = child.getValue(Acreedor.class);
+    private void cargarAcreedor() {
+        if(isNetworkAvailable()){
+            database = FirebaseDatabase.getInstance().getReference().child("Acreedores");
+            database.addValueEventListener(new ValueEventListener() {
+                @Override
+                public void onDataChange(@NonNull DataSnapshot snapshot) {
+                    for (DataSnapshot child : snapshot.getChildren()){
+                        if(child.child("nif").getValue().toString().equals(nif)){
+                            acreedor = child.getValue(Acreedor.class);
 
-                        Glide.with(getApplicationContext()).load(acreedor.getImageUri()).into(ivAcreedor);
-                        tvNombre.setText(acreedor.getNombre());
-                        tvNif.setText(acreedor.getNif());
-                        tvEmail.setText(acreedor.getEmail());
-                        tvTelefono.setText(acreedor.getTelefono());
+                            Glide.with(getApplicationContext()).load(acreedor.getImageUri()).into(ivAcreedor);
+                            tvNombre.setText(acreedor.getNombre());
+                            tvNif.setText(acreedor.getNif());
+                            tvEmail.setText(acreedor.getEmail());
+                            tvTelefono.setText(acreedor.getTelefono());
+                        }
                     }
+                    cargarToolbar();
                 }
-                cargarToolbar();
-            }
 
-            @Override
-            public void onCancelled(@NonNull DatabaseError error) {
+                @Override
+                public void onCancelled(@NonNull DatabaseError error) {
 
-            }
-        });
+                }
+            });
+        }else{
+            SQLAcreedoresController sql = new SQLAcreedoresController(getApplicationContext());
+            acreedor = sql.getAcreedor(nif);
+            tvNombre.setText(acreedor.getNombre());
+            tvNif.setText(acreedor.getNif());
+            tvEmail.setText(acreedor.getEmail());
+            tvTelefono.setText(acreedor.getTelefono());
+            cargarToolbar();
+        }
     }
 
     @Override
@@ -105,12 +115,35 @@ public class VistaAcreedor extends AppCompatActivity {
 
         } else if (id == R.id.menuBorrar) {
 
+            if(acreedor.getNombre().isEmpty()){
+                acreedor.setNombre("nombreDefault");
+            }
+            if(acreedor.getEmail().isEmpty()){
+                acreedor.setEmail("emailDefault");
+            }
+            if(acreedor.getTelefono().isEmpty()){
+                acreedor.setTelefono("telefonoDefault");
+            }
+
             EditarAcreedor.aux=true;
-            database = FirebaseDatabase.getInstance().getReference().child("Acreedores").child(acreedor.getNif());
-            database.removeValue();
+            SQLAcreedoresController sql = new SQLAcreedoresController(getApplicationContext());
+            if(isNetworkAvailable()){
+                database = FirebaseDatabase.getInstance().getReference().child("Acreedores").child(acreedor.getNif());
+                database.removeValue();
+            }else{
+                sql.borrarAcreedorAux(acreedor);
+            }
+            sql.borrarAcreedor(nif);
             finish();
 
         }
         return super.onOptionsItemSelected(item);
+    }
+
+    private boolean isNetworkAvailable() {
+        ConnectivityManager connectivityManager
+                = (ConnectivityManager) getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo activeNetworkInfo = connectivityManager.getActiveNetworkInfo();
+        return activeNetworkInfo != null && activeNetworkInfo.isConnected();
     }
 }
